requirements.txt
streamlit>=1.32.0
requests>=2.31.0
Pillow>=10.0.0
numpy>=1.26.0
pandas>=2.1.0
sqlalchemy>=2.0.0
